+++ UnderAutomation.Fanuc.ObfuscatedSources/Ftp/Diagnosis/Feature.cs
@@ -1,0 +1,26 @@
+//              WARNING
+// This file is an empty shell containing only public C# items.
+// The internal code is hidden; to access it, you need to obtain a Source licence of the library.
+
+
+namespace Ftp.Diagnosis {
+
+	public class Feature {
+
+
+		public Feature()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+		}
+
+		/// <summary>
+		/// Name of the feature
+		/// </summary>
+		public string Name { get; set; }
+
+		/// <summary>
+		/// Order number of the feature (set of 4 alphanumeric characters)
+		/// </summary>
+		public string OrderNo { get; set; }
+	}
+}