+++ UnderAutomation.Fanuc.ObfuscatedSources/Ftp/Diagnosis/Features.cs
@@ -1,0 +1,27 @@
+//              WARNING
+// This file is an empty shell containing only public C# items.
+// The internal code is hidden; to access it, you need to obtain a Source licence of the library.
+
+
+namespace Ftp.Diagnosis {
+
+	public class Features {
+
+
+		public Features()
+		{
+			// Source is hidden, a Source licence is needed to access internal code...
+		}
+
+		/// <summary>
+		/// List of features
+		/// </summary>
+		public Feature[] FeaturesList { get; }
+
+
+		public bool HasTelnet { get; }
+
+
+		public bool HasSnpx { get; }
+	}
+}